+++ Scripts/AdsManager/IAdsNetworkHelper.cs
@@ -14,4 +14,6 @@ public interface IAdsNetworkHelper
     void RequestInterstitialRewardedNoShow(AdPlacement.Type placementType, RewardDelegate onAdLoaded = null);
     void ShowInterstitialRewarded(AdPlacement.Type placementType, RewardDelegate onAdClosed);
     void Reward(AdPlacement.Type placementType, RewardDelegate onFinish);
+    void RequestAppOpenAd(AdPlacement.Type placementType, AdsManager.InterstitialDelegate onAdLoaded = null);
+    void ShowAppOpenAd(AdPlacement.Type placementType, AdsManager.InterstitialDelegate onAdClosed = null);
 }